[Credentialless]: Make fetch.tentative.https.html more reliable.
Previously, the test was assuming the window can be closed immediately
after sending the last fetch request.
This is not true, because closing the window might interrupt previously
scheduled fetch requests and fails the test.
We need to wait for the test completion before assuming we can close the
windows.

I did a local test with 100 wpt run:
- without fix: 16/100 failures.
- with fix: 0/100 failures.

Bug:1175099,1198255
Change-Id: I70a56ba4821d46d12f45c38d8850c876c5957649