[LayoutNG] OOF positioned offset in a nested multicol
OOF positioned elements (abspos or fixedpos) in a nested multicol
could get the wrong offset if the containing block started in a column
that was not the first in a given multicol fragment. The reason being
that we had adjusted the containing block offset in CL:2680070 to take
into account columns from previous fragments. However, if the containing
block was not in the first column in a multicol fragment, we would have
already adjusted for this in NGContainerFragmentBuilder (given that its
column's builder would have access to the previous column break token).

To fix this, utilize the previous_break_token_ stored in
NGBoxFragmentBuilder (which is available in both cases above) to
determine the correct offset for the containing block. This is now
updated via NGBoxFragmentBuilder::AdjustOffsetForFragmentainer().

Bug: 1198578
Change-Id: I3c183dd83ed45e240fb9f62347dd388e370b1510